Adds a deprecation warning for drop-down-icon
This component really should have had a deprecation warning from 0.14.0 when the old `./menu/menu` implementation was deprecated. This component is not used in any other components in src or docs. It also is not documented on the site. I think we should add this warning ASAP and plan for removal along with `./menu/menu` in our first `0.15.0` release candidate.