Add `mergeDateAndTime` to the utils and use it from <Calendar/>
Currently, the Calendar.onDateSelect() handler simply calls setHours() and .setMinutes() to carry over the time information from the previous value to the new date-only value selected in the Calendar. This commit adds a function `mergeDateAndTime` to the date utils, which defaults to this behavior, but allows the user to customize how the time information is carried over. One use case is carrying over the seconds in addition. Another use case is values that do not have any time information, in which case there is nothing to carry over.