[SpeedDial] Fix navigation between SpeedDialActions (#12725)
Closes #12361
Addresses #12159

* [SpeedDialAction] Fix react warnings

className and classes.root was not applied to anything or forwarded to
anything in the past. They should be implemented in the future however.

A missing tooltipTitle caused a warning further down the component tree.

* [SpeedDial] Fix react warnings

onKeyDown was applied to the button and the SpeedDialAction. However the
Action did not process onKeyDown and passing the same
event listener to different components is ambiguous.

* [SpeedDial] Fix keyboard navigation when focused

* [SpeedDial] Fix onKeyDown not being called

* [SpeedDial] Implement navigation behavior

Following the proposed spec with alternative B

* [SpeedDial] Fix buttonRef override

* [SpeedDial] Add test cases for navigation

* [docs] Remove unresolved merge residues

* [SpeedDial] Fix onKeyDown not called due to merge conflicts

* [SpeedDial] switch from wrapping to clamping in keyboard navigation

- FAB is included in navigation
- Fixed potential issues when mixing keyboard and mouse navigation or
  refocusing a SpeedDial

* [SpeedDial] Support createRef in ButtonProps